pyproject.toml
[project]
name = "tnega"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "pydantic-ai>=1.7.0",
    "loguru>=0.7.0",
    "pytest>=8.0.0",
    "pandas>=2.2.0",
    "tweepy>=4.16.0",
    "aiohttp>=3.9.0",
    "async-lru>=2.0.0",
    "oauthlib>=3.2.0",
    "python-dotenv>=1.0.0",
    "pydantic-settings>=2.11.0",
    "httpx>=0.28.1",
    "aiofiles>=25.1.0",
    "openpyxl>=3.1.5",
]
